+++ Marlin/stepper.h
@@ -84,11 +84,11 @@ extern block_t *current_block;  // A pointer to the block currently being traced
 
 void quickStop();
 
-void digitalPotWrite(int address, int value);
+void digitalPotWrite(int16_t address, int16_t value);
 void microstep_ms(uint8_t driver, int8_t ms1, int8_t ms2);
 void microstep_mode(uint8_t driver, uint8_t stepping);
 void digipot_init();
-void digipot_current(uint8_t driver, int current);
+void digipot_current(uint8_t driver, int16_t current);
 void microstep_init();
 void microstep_readings();
